feat: adapt create_term and create_and_name_term actions to work when a user has their toggleterm config's direction key set to 'float'
If the term direction is 'float' and exit_on_action is false, create the terminal in the background (term:spawn) to prevent flashes that are caused by telescope and toggleterm both being floating windows. If exit_on_action is true, treat it the same as 'horizontal' and 'vertical' toggleterm terminals, which is to exit telescope and focus the terminal.