fix: check what the user's telescope config option of 'defaults.sorting_strategy' is when calling 'refresh_picker'/'create_finder'
When 'defaults.sorting_strategy = descending', telescope row numbers begin at 250 instead of 0. This was causing 'create_finder' to return the wrong 'new_row_number' value to 'refresh_picker' since it was only capable of handling 0-based row numbers, which is how telescope numbers rows when 'defaults.sorting_strategy = ascending'.

Similarly, additional logic needed to be added directly to 'refresh_picker' for actions that pass 'deleted = true' to it so that it can also handle when 'defaults.sorting_strategy = descending'.